WIP - First pass at refining ReusableView
ReusableView is a useful protocol, but we can refine it in two ways:

- View controllers aren't reusable but often need an identifier, similar to reuseIdentifier, in order to instantiate from a storyboard
- Cells are reusable, but may have other common needs like estimated size

This makes Identifiable a base protocol, and ReusableView inherits from it, and ReusableCell inherits from that. UIKit and MapKit types are extended to conform to these protocols as appropriate. Cell types _don't_ conform to ReusableCell by default because it's better to require end-user cell types to specify an estimated size than setting a default like CGSize.zero. Helper methods have been updated to make it easier to register, dequeue or instantiate these views and cells in a type safe way.

NibLoadableView now also inherits from Identifiable, and another static var was added that finds a views XIB in its bundle.

Todo:

- Add documentation
- Update tests
- Rebase on Swift 4.2 compatible branch to make sure everything runs and passes tests